Add Event Log viewer and Noise Analysis features
- Add WmiEventLogService for remote Sysmon event log queries
- Add NoiseAnalysisService with role-based thresholds and exclusion generation
- Add EventsController and AnalysisController API endpoints
- Add EventsPage and NoiseAnalysisPage frontend components
- Add DeploymentQueuePanel and XmlEditor components
- Add Theme and DeploymentQueue context providers
- Update .gitignore to exclude local dev files

+++ src/SysmonConfigPusher.Core/Models/NoiseAnalysis.cs
@@ -5,7 +5,7 @@ public class NoiseAnalysisRun
     public int Id { get; set; }
     public int ComputerId { get; set; }
     public Computer Computer { get; set; } = null!;
-    public int TimeRangeHours { get; set; }
+    public double TimeRangeHours { get; set; }
     public int TotalEvents { get; set; }
     public DateTime AnalyzedAt { get; set; } = DateTime.UtcNow;
